% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scion_inputs.R
\name{detect_scion_format}
\alias{detect_scion_format}
\title{Detect whether a file is GCT(x) or delimited-text, by extension}
\usage{
detect_scion_format(path)
}
\description{
Detect whether a file is GCT(x) or delimited-text, by extension
}
\keyword{internal}
