% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_network.R
\name{infer_hub_network}
\alias{infer_hub_network}
\title{Infer the network connecting cluster hub genes}
\usage{
infer_hub_network(
  target_data,
  reg_data,
  myhubs,
  weightthreshold,
  normalize,
  num.cores,
  engine,
  ptm_sep,
  seed,
  ...
)
}
\description{
Infer the network connecting cluster hub genes
}
\keyword{internal}
