% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_network.R
\name{infer_network_clustered}
\alias{infer_network_clustered}
\title{Infer one network per cluster, optionally connecting cluster hubs}
\usage{
infer_network_clustered(
  target_data,
  reg_data,
  cluster_assignment,
  weightthreshold,
  normalize,
  connect_hubs,
  num.cores,
  engine,
  ptm_sep,
  ...
)
}
\description{
Infer one network per cluster, optionally connecting cluster hubs
}
\keyword{internal}
