% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scion.R
\name{write_scion_network}
\alias{write_scion_network}
\title{Write a SCION network to a Cytoscape-importable file}
\usage{
write_scion_network(network, output_file)
}
\arguments{
\item{network}{an edge table as returned in the \code{network} element of
\code{\link[=run_scion]{run_scion()}}'s result.}

\item{output_file}{path to write to (tab-separated).}
}
\description{
Write a SCION network to a Cytoscape-importable file
}
