% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_network.R
\name{infer_network_single}
\alias{infer_network_single}
\title{Infer a single (non-clustered) network}
\usage{
infer_network_single(
  target_data,
  reg_data,
  weightthreshold,
  normalize,
  num.cores,
  engine,
  ...
)
}
\description{
Infer a single (non-clustered) network
}
\keyword{internal}
