% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_network.R
\name{pick_hub_genes}
\alias{pick_hub_genes}
\title{Pick the hub gene(s) (highest out-degree regulator) from an edge table}
\usage{
pick_hub_genes(edge_table)
}
\arguments{
\item{edge_table}{a data frame as returned by \code{\link[=weight_matrix_to_edges]{weight_matrix_to_edges()}}.}
}
\value{
a character vector of regulator names with the maximum out-degree
(length > 1 if there is a tie).
}
\description{
Pick the hub gene(s) (highest out-degree regulator) from an edge table
}
\keyword{internal}
