% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_diagnostics.R
\name{plot_outdegree_distribution}
\alias{plot_outdegree_distribution}
\title{Plot the distribution of regulator out-degrees in a network}
\usage{
plot_outdegree_distribution(network, bins = 30)
}
\arguments{
\item{network}{an edge table with a \code{Regulator} column.}

\item{bins}{number of histogram bins.}
}
\value{
a \code{ggplot2} object.
}
\description{
Plot the distribution of regulator out-degrees in a network
}
